pyproject.toml
[project]
name = "inquiry-automation-pipeline"
version = "0.1.0"
description = "Intelligent Automation Pipeline with MLOps Governance for Client Inquiry Classification and Routing"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.9",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "sentencepiece>=0.1.99",
    "scikit-learn>=1.3.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "apache-airflow>=2.7.0",
    "apache-airflow-providers-postgres>=5.7.0",
    "mlflow>=2.8.0",
    "prometheus-client>=0.19.0",
    "streamlit>=1.28.0",
    "plotly>=5.18.0",
    "boto3>=1.29.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "faker>=20.0.0",
    "pyyaml>=6.0.1",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.3",
    "redis>=5.0.0",
    "celery>=5.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.5",
    "mypy>=1.7.0",
    "locust>=2.17.0",
]
terraform = [
    "boto3>=1.29.0",
    "python-terraform>=0.10.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

